pkg: move files from src to lib.
