+++ lib/bsigner.js
@@ -2,18 +2,18 @@
 
 // import stuff from src/app to export
 
-const {Hardware} = require('../src/hardware');
-const {Path} = require('../src/path');
+const {DeviceManager} = require('../lib/device/manager');
+const {Path} = require('../lib/path');
 const {
   prepareSign,
   generateToken,
   prepareSignMultisig,
   guessPath,
   getKnownPaths
-} = require('../src/app');
+} = require('../lib/app');
 
 // classes
-exports.Hardware = Hardware;
+exports.DeviceManager = DeviceManager;
 exports.Path = Path;
 
 // app methods
