+++ lib/common.js
@@ -13,13 +13,10 @@ const network = require('bcoin/lib/protocol/networks');
 
 /*
  * vendors to act as backends
- * for the Hardware signer
+ * for the Hardware signer.
  */
 const vendors = {
-  LEDGER: 'LEDGER',
-  TREZOR: 'TREZOR',
-  LOCAL: 'LOCAL',
-  ANY: 'ANY'
+  LEDGER: 'LEDGER'
 };
 
 const bip44 = {
@@ -106,7 +103,7 @@ function parsePath(path, hard) {
   for (let i = 1; i < parts.length; i++) {
     let part = parts[i];
 
-    const last = part[part.length - 1]
+    const last = part[part.length - 1];
     const hardened = last === '\'' || last === 'h';
 
     if (hardened)
