+++ lib/path.js
